// apps/docs-app/app/builder/shapes/file-builder/generated-file-input.shape.ts
export interface GeneratedFileInputShape {
  featureCellImports: string;
  initialValueDisplay: string | undefined;
  featureCellBehaviorArray: string;
  serviceVaultName: string;
  serviceName: string;
  featureCellKey: string;
  featureCellControllerArray: string;
  vaultImports: string;
  npm: string;
  type: string;
  vaultChain: string;
  aiAssist: string;
  coreBehaviorNotes: string;
  serviceExamples: string;
  componentExamples: string;
  instantiatedServiceName: string;
  interfaceDefinition: string;
}